feat: improve grpc exception handling in cart-service and add Grpc security filters in cart-sarvice and payment-service, change stockQuantity field type in Product entity from int to long

+++ services/product-service/src/main/java/com/productservice/dto/ProductReservedQuantity.java
@@ -4,6 +4,6 @@ import java.util.UUID;
 
 public record ProductReservedQuantity(
         UUID productId,
-        Integer reservedQuantity
+        Long reservedQuantity
 ) {
 }
