+++ services/product-service/src/main/java/com/productservice/dto/request/CreateProductRequest.java
@@ -27,5 +27,5 @@ public record CreateProductRequest(
 
         @NotNull
         @Min(0)
-        Integer stockQuantity
+        Long stockQuantity
 ) {}
